Cast fixes

+++ opcodeFunc.c
@@ -441,7 +441,7 @@ void setConditionals(int des, u_int64_t *regArr, unsigned int *condFlag) {
         case 3:
           //DEBUG PRINT
           printf("At reg%d, value is %d", des, regArr[des]);
-          if(regArr[des] < 0) {
+          if((int64_t)regArr[des] < 0) {
             condFlag[i] = 1;
           } else {
             condFlag[i] = 0;
